Upgrade to Storybook v7 with Vite
