+++ .changeset/shiny-pans-sit.md
@@ -1,0 +1,5 @@
+---
+'@shopify/polaris': minor
+---
+
+Upgrade to Storybook v7 with Vite
