+++ polaris-react/.storybook/addons/global-controls-panel/preset.js
@@ -1,0 +1,6 @@
+function managerEntries(entry = []) {
+  // Unfortunately have to do this dance for Storybook v7 to load our addon
+  return [...entry, require.resolve('./manager')];
+}
+
+export default {managerEntries};
