+++ polaris-react/postcss.config.mjs
@@ -1,0 +1,8 @@
+import postcssPlugins from './config/postcss-plugins.js';
+
+const config = {
+  plugins: postcssPlugins,
+};
+
+// eslint-disable-next-line import/no-default-export
+export default config;
